+++ scripts/docs.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-./py3/bin/sphinx-build docs/source/ docs/html