+++ scripts/karma.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-clear
-node_modules/karma/bin/karma start js/karma.conf.js