+++ scripts/rollup.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-node_modules/rollup/dist/bin/rollup --config js/rollup.conf.js "$@"