+++ src/cone/app/tests/test_browser_resources.py
@@ -130,7 +130,7 @@ class TestBrowserResources(TileTestCase):
 
         self.assertTrue(styles[0].directory.endswith(np('/static/cone')))
         self.assertEqual(styles[0].path, 'cone')
-        self.assertEqual(styles[0].file_name, 'cone.app.css')
+        self.assertEqual(styles[0].file_name, 'cone.app.min.css')
         self.assertTrue(os.path.exists(styles[0].file_path))
 
         self.assertTrue(styles[1].directory.endswith(np('/static/cone')))